+++ src/components/Modal/CandidatesList.tsx
